requirements.txt
Django==2.2.6
djangorestframework==3.10.3
mysqlclient
pytz==2019.3
sqlparse==0.3.0
